+++ CodeStarterCamp_Week1/main.swift
@@ -8,20 +8,6 @@
 
 import Foundation
 
-func orderIceCream(width: Int, height: Int, stickSize: Int = 4) {
-    makeIceCream(width: width, height: height)
-    makeStick(width: width, stickSize: stickSize)
-}
-
-func makeIceCream(width: Int, height: Int) {
-    for _ in 0 ..< height {
-        print(String(repeating: "*", count: width))
-    }
-}
-
-func makeStick(width: Int, stickSize: Int) {
-    print(String(repeating: " ", count: (width-3)/2) + "| |")
-}
-
+// MARK: - [STEP 2] 아이스크림 만들기
 orderIceCream(width: 10, height: 8)
